LayoutTests/imported/w3c/web-platform-tests/navigation-api/updateCurrentEntry-method/same-document-away-and-back-location-api-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: null is not an object (evaluating 'entry0.index')

Harness Error (FAIL), message = Unhandled rejection: null is not an object (evaluating 'entry0.index')

TIMEOUT entry.getState() behavior after navigating away using the location API, then back Test timed out

Harness Error (FAIL), message = Unhandled rejection: null is not an object (evaluating 'entry0.index')

TIMEOUT entry.getState() behavior after navigating away using the location API, then back Test timed out
